Extract AuthorClass
